# specs/001-desktop-career-agent/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CareerOS Local Desktop Additions
  version: 1.3.0
  description: Local model, application workflow and recovery interfaces for feature 001.
servers:
  - url: http://127.0.0.1:{port}/api/v1
    variables:
      port:
        default: "8000"
security:
  - desktopSession: []
paths:
...
  /applications/agenda:
    get:
      operationId: getApplicationAgenda
      description: Builds a deterministic private queue from owned scalar application projections.
      security:
        - desktopSession: []
          userAccessToken: []
      parameters:
        - name: local_day_end
          in: query
          required: true
          description: Timezone-aware instant of the browser's next local midnight, at most 26 hours after generation.
          schema: { type: string, format: date-time }
        - name: horizon_days
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 30, default: 7 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: User-scoped agenda with explicit later and limit-truncation counts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationAgendaResponse" }
        "401": { description: The local user session is not authenticated. }
        "422": { description: A query bound or persisted next-action projection is invalid. }
...
  /applications/{application_id}/readiness:
    get:
      operationId: getApplicationReadiness
      security:
        - desktopSession: []
          userAccessToken: []
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Deterministic local preflight completeness report.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationReadinessReport" }
        "404": { description: The application does not belong to the current user. }
  /applications/{application_id}/readiness/export:
    get:
      operationId: exportApplicationReadiness
      security:
        - desktopSession: []
          userAccessToken: []
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: format
          in: query
          required: false
          schema: { type: string, enum: [json, markdown], default: json }
      responses:
        "200":
          description: Canonical report bytes; X-Content-SHA256 hashes the exact response body.
          headers:
            X-Content-SHA256:
              schema: { type: string, pattern: "^[a-f0-9]{64}$" }
            Content-Disposition:
              schema: { type: string }
          content:
            application/json:
              schema: { type: string, format: binary }
            text/markdown:
              schema: { type: string, format: binary }
        "404": { description: The application does not belong to the current user. }
  /applications/{application_id}/preparation:
    patch:
      operationId: updateApplicationPreparation
      security:
        - desktopSession: []
          userAccessToken: []
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationPreparationUpdate" }
      responses:
        "200": { description: Revisioned application with a content-free preparation audit event. }
        "404": { description: The application does not belong to the current user. }
        "409": { description: The expected revision is stale. }
        "422": { description: A URL, email, resume link or update payload is invalid. }
  /applications/{application_id}/dossier-draft:
    parameters:
      - name: application_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getApplicationDossierDraft
      security:
        - desktopSession: []
          userAccessToken: []
      responses:
        "200":
          description: The owned SQLite-backed working draft, or null when no draft exists.
          headers:
            Cache-Control:
              schema: { const: "private, no-store" }
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: "#/components/schemas/ApplicationDossierDraftResponse" }
                  - { type: "null" }
        "401": { description: The local user session is not authenticated. }
        "404": { description: The application does not belong to the current user. }
    put:
      operationId: putApplicationDossierDraft
      description: Creates or compare-and-swap updates one bounded working draft.
      security:
        - desktopSession: []
          userAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationDossierDraftPut" }
      responses:
        "200":
          description: The stored draft with its next compare-and-swap revision.
          headers:
            Cache-Control:
              schema: { const: "private, no-store" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApplicationDossierDraftResponse" }
        "401": { description: The local user session is not authenticated. }
        "404": { description: The application does not belong to the current user. }
        "409":
          description: The application, linked resume or working draft revision is stale.
        "422": { description: The bounded draft payload is invalid. }
    delete:
      operationId: deleteApplicationDossierDraft
      security:
        - desktopSession: []
          userAccessToken: []
      parameters:
        - name: expected_revision
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "204": { description: The exact saved draft revision was deleted. }
        "401": { description: The local user session is not authenticated. }
        "404": { description: The application or draft does not exist for the current user. }
        "409": { description: The working draft revision is stale. }
        "422": { description: The expected revision is invalid. }
  /applications/{application_id}/dossiers:
    post:
      operationId: publishApplicationDossier
      description: >
        Publishes directly when no working draft exists. If a draft exists, the request must name
        and exactly match its current revision; that row is consumed in the immutable event
        transaction.
      security:
        - desktopSession: []
          userAccessToken: []
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ApplicationDossierCreate" }
      responses:
        "201":
          description: The immutable dossier event and updated application revision.
        "401": { description: The local user session is not authenticated. }
        "404": { description: The application does not belong to the current user. }
        "409":
          description: The application or saved draft changed before publication.
        "422":
          description: Readiness, evidence, content or bounded artifact validation failed.
components:
  securitySchemes:
    desktopSession:
      type: apiKey
      in: header
      name: X-CareerOS-Session
      description: Per-launch secret supplied by the native shell; disabled in explicit browser development.
    userAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authenticated local CareerOS user access token.
    refreshCookie:
      type: apiKey
      in: cookie
      name: careeros_refresh_token
      description: HttpOnly, path-scoped, single-use JWT bound to one persisted session family.
...
  schemas:
...
    ApplicationAgendaResponse:
      type: object
      additionalProperties: false
      required: [generated_at, local_day_end, horizon_end, active_count, visible_count, later_count, truncated_count, items]
      properties:
        generated_at: { type: string, format: date-time }
        local_day_end: { type: string, format: date-time }
        horizon_end: { type: string, format: date-time }
        active_count: { type: integer, minimum: 0 }
        visible_count: { type: integer, minimum: 0 }
        later_count: { type: integer, minimum: 0 }
        truncated_count: { type: integer, minimum: 0 }
        items:
          type: array
          maxItems: 200
          items: { $ref: "#/components/schemas/ApplicationAgendaItem" }
    ApplicationAgendaItem:
      type: object
      additionalProperties: false
      required: [application_id, application_revision, title, company, current_stage, latest_event_at, state, next_action]
      properties:
        application_id: { type: string, format: uuid }
        application_revision: { type: integer, minimum: 1 }
        title: { type: string, minLength: 1, maxLength: 240 }
        company: { type: string, minLength: 1, maxLength: 240 }
        current_stage:
          type: string
          enum: [saved, preparing, applied, screening, interview, offer, accepted]
        latest_event_at: { type: string, format: date-time }
        state:
          type: string
          enum: [overdue, today, upcoming, unscheduled, needs_action]
        next_action:
          oneOf:
            - { $ref: "#/components/schemas/ApplicationNextAction" }
            - { type: "null" }
    ApplicationNextAction:
      type: object
      additionalProperties: false
      required: [id, title, due_at, priority]
      properties:
        id: { type: string, minLength: 1, maxLength: 36 }
        title: { type: string, minLength: 1, maxLength: 500 }
        due_at:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
        priority: { type: string, enum: [low, normal, high, urgent] }
    ApplicationPreparationUpdate:
      type: object
      additionalProperties: false
      required: [expected_revision]
      anyOf:
        - required: [title]
        - required: [company]
        - required: [description]
        - required: [application_url]
        - required: [application_email]
        - required: [resume_version_id]
      properties:
        expected_revision: { type: integer, minimum: 1 }
        title: { type: [string, "null"], maxLength: 240 }
        company: { type: [string, "null"], maxLength: 240 }
        description: { type: [string, "null"], maxLength: 100000 }
        application_url: { type: [string, "null"], maxLength: 2048 }
        application_email: { type: [string, "null"], maxLength: 320 }
        resume_version_id: { type: [string, "null"], maxLength: 36 }
    DossierDraftAnswer:
      type: object
      additionalProperties: false
      required: [client_id]
      properties:
        client_id: { type: string, minLength: 1, maxLength: 100 }
        question: { type: string, maxLength: 1000, default: "" }
        answer: { type: string, maxLength: 20000, default: "" }
    DossierDraftChecklistItem:
      type: object
      additionalProperties: false
      required: [client_id]
      properties:
        client_id: { type: string, minLength: 1, maxLength: 100 }
        label: { type: string, maxLength: 500, default: "" }
        completed: { type: boolean, default: false }
    DossierDraftRequirementEvidence:
      type: object
      additionalProperties: false
      required: [client_id]
      properties:
        client_id: { type: string, minLength: 1, maxLength: 100 }
        requirement: { type: string, maxLength: 2000, default: "" }
        evidence_fact_ids:
          type: array
          maxItems: 10
          uniqueItems: true
          items: { type: string, format: uuid }
    ApplicationDossierDraftContent:
      type: object
      additionalProperties: false
      required: [requirement_matrix]
      description: Incomplete draft rows are allowed; client_id values are unique per collection.
      properties:
        cover_letter: { type: [string, "null"], maxLength: 30000 }
        answers:
          type: array
          maxItems: 25
          items: { $ref: "#/components/schemas/DossierDraftAnswer" }
        checklist:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/DossierDraftChecklistItem" }
        requirement_matrix:
          type: array
          minItems: 1
          maxItems: 25
          items: { $ref: "#/components/schemas/DossierDraftRequirementEvidence" }
    ApplicationDossierDraftPut:
      type: object
      additionalProperties: false
      required:
        - expected_application_revision
        - resume_version_id
        - content
      properties:
        expected_revision: { type: [integer, "null"], minimum: 1 }
        expected_application_revision: { type: integer, minimum: 1 }
        resume_version_id: { type: string, format: uuid }
        content: { $ref: "#/components/schemas/ApplicationDossierDraftContent" }
    ApplicationDossierDraftResponse:
      type: object
      additionalProperties: false
      required:
        - application_id
        - resume_version_id
        - application_revision
        - revision
        - content
        - created_at
        - updated_at
      properties:
        application_id: { type: string, format: uuid }
        resume_version_id: { type: string, format: uuid }
        application_revision: { type: integer, minimum: 1 }
        revision: { type: integer, minimum: 1 }
        content: { $ref: "#/components/schemas/ApplicationDossierDraftContent" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DossierAnswer:
      type: object
      additionalProperties: false
      required: [question, answer]
      properties:
        question: { type: string, minLength: 1, maxLength: 1000 }
        answer: { type: string, minLength: 1, maxLength: 20000 }
    DossierChecklistItem:
      type: object
      additionalProperties: false
      required: [label]
      properties:
        label: { type: string, minLength: 1, maxLength: 500 }
        completed: { type: boolean, default: false }
    DossierRequirementEvidence:
      type: object
      additionalProperties: false
      required: [requirement, evidence_fact_ids]
      properties:
        requirement: { type: string, minLength: 1, maxLength: 2000 }
        evidence_fact_ids:
          type: array
          minItems: 1
          maxItems: 10
          uniqueItems: true
          items: { type: string, format: uuid }
    ApplicationDossierCreate:
      type: object
      additionalProperties: false
      required: [expected_revision, requirement_matrix]
      properties:
        expected_revision: { type: integer, minimum: 1 }
        expected_draft_revision: { type: [integer, "null"], minimum: 1 }
        cover_letter: { type: [string, "null"], maxLength: 30000 }
        answers:
          type: array
          maxItems: 25
          items: { $ref: "#/components/schemas/DossierAnswer" }
        checklist:
          type: array
          maxItems: 50
          items: { $ref: "#/components/schemas/DossierChecklistItem" }
        requirement_matrix:
          type: array
          minItems: 1
          maxItems: 25
          items: { $ref: "#/components/schemas/DossierRequirementEvidence" }
    ApplicationReadinessReport:
      type: object
      additionalProperties: false
      required: [schema_version, application_id, application_revision, role_title, company, status, score_kind, completeness_score, blocker_count, warning_count, checks, fingerprint]
      properties:
        schema_version: { const: "1.0" }
        application_id: { type: string, format: uuid }
        application_revision: { type: integer, minimum: 1 }
        role_title: { type: string }
        company: { type: string }
        status: { type: string, enum: [ready, action_needed, blocked] }
        score_kind: { const: preflight_completeness }
        completeness_score: { type: integer, minimum: 0, maximum: 100 }
        blocker_count: { type: integer, minimum: 0 }
        warning_count: { type: integer, minimum: 0 }
        checks:
          type: array
          items: { $ref: "#/components/schemas/ApplicationReadinessCheck" }
        fingerprint: { type: string, pattern: "^[a-f0-9]{64}$" }
    ApplicationReadinessCheck:
      type: object
      additionalProperties: false
      required: [id, status, points_awarded, points_available, evidence]
      properties:
        id: { type: string }
        status: { type: string, enum: [pass, warning, blocker] }
        points_awarded: { type: integer, minimum: 0, maximum: 100 }
        points_available: { type: integer, minimum: 1, maximum: 100 }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/ReadinessEvidence" }
        action: { type: [string, "null"] }
    ReadinessEvidence:
      type: object
      additionalProperties: false
      required: [key, value]
      properties:
        key: { type: string }
        value: { type: string }